requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.4.1
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
Google-Images-Search==1.3.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
Pillow==7.1.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
python-resize-image==1.1.19
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
termcolor==1.1.0
udpy==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
yarl==1.5.1
